TIP-706-revert: Move files and invert locales and channels
