+++ src/Pim/Bundle/CatalogBundle/tests/integration/ElasticSearch/IndexConfiguration/PimCatalogTextAreaIntegration.php
@@ -21,7 +21,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'bool' => [
                     'filter' => [
                         'query_string' => [
-                            'default_field' => 'values.description-text.<all_locales>.<all_channels>',
+                            'default_field' => 'values.description-text.<all_channels>.<all_locales>',
                             'query'         => 'an*',
                         ],
                     ],
@@ -41,7 +41,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'bool' => [
                     'filter' => [
                         'query_string' => [
-                            'default_field' => 'values.description-text.<all_locales>.<all_channels>.raw',
+                            'default_field' => 'values.description-text.<all_channels>.<all_locales>.raw',
                             'query'         => '*My*',
                         ],
                     ],
@@ -61,12 +61,12 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'bool' => [
                     'must_not' => [
                         'query_string' => [
-                            'default_field' => 'values.description-text.<all_locales>.<all_channels>.raw',
+                            'default_field' => 'values.description-text.<all_channels>.<all_locales>.raw',
                             'query'         => '*cool\\ product*',
                         ],
                     ],
                     'filter'   => [
-                        'exists' => ['field' => 'values.description-text.<all_locales>.<all_channels>'],
+                        'exists' => ['field' => 'values.description-text.<all_channels>.<all_locales>'],
                     ],
                 ],
             ],
@@ -84,7 +84,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                     'bool' => [
                         'filter' => [
                             'term' => [
-                                'values.description-text.<all_locales>.<all_channels>.raw' => 'yeah, love description',
+                                'values.description-text.<all_channels>.<all_locales>.raw' => 'yeah, love description',
                             ],
                         ],
                     ],
@@ -103,11 +103,11 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                     'bool' => [
                         'must_not' => [
                             'term' => [
-                                'values.description-text.<all_locales>.<all_channels>.raw' => 'yeah, love description',
+                                'values.description-text.<all_channels>.<all_locales>.raw' => 'yeah, love description',
                             ],
                         ],
                         'filter'   => [
-                            'exists' => ['field' => 'values.description-text.<all_locales>.<all_channels>.raw'],
+                            'exists' => ['field' => 'values.description-text.<all_channels>.<all_locales>.raw'],
                         ],
                     ],
                 ],
@@ -124,7 +124,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'query' => [
                     'bool' => [
                         'must_not' => [
-                            'exists' => ['field' => 'values.description-text.<all_locales>.<all_channels>'],
+                            'exists' => ['field' => 'values.description-text.<all_channels>.<all_locales>'],
                         ],
                     ],
                 ],
@@ -141,7 +141,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'query' => [
                     'bool' => [
                         'filter' => [
-                            'exists' => ['field' => 'values.description-text.<all_locales>.<all_channels>'],
+                            'exists' => ['field' => 'values.description-text.<all_channels>.<all_locales>'],
                         ],
                     ],
                 ],
@@ -160,7 +160,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
             ],
             'sort'  => [
                 [
-                    'values.description-text.<all_locales>.<all_channels>.raw' => [
+                    'values.description-text.<all_channels>.<all_locales>.raw' => [
                         'order'   => 'asc',
                         'missing' => '_first',
                     ],
@@ -184,7 +184,7 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
             ],
             'sort'  => [
                 [
-                    'values.description-text.<all_locales>.<all_channels>.raw' => [
+                    'values.description-text.<all_channels>.<all_locales>.raw' => [
                         'order'   => 'desc',
                         'missing' => '_last',
                     ],
@@ -210,8 +210,8 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'identifier' => 'product_1',
                 'values'     => [
                     'description-text' => [
-                        '<all_locales>' => [
-                            '<all_channels>' => 'My product description',
+                        '<all_channels>' => [
+                            '<all_locales>' => 'My product description',
                         ],
                     ],
                 ],
@@ -220,8 +220,8 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'identifier' => 'product_2',
                 'values'     => [
                     'description-text' => [
-                        '<all_locales>' => [
-                            '<all_channels>' => 'Another cool product, great !',
+                        '<all_channels>' => [
+                            '<all_locales>' => 'Another cool product, great !',
                         ],
                     ],
                 ],
@@ -230,8 +230,8 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'identifier' => 'product_3',
                 'values'     => [
                     'description-text' => [
-                        '<all_locales>' => [
-                            '<all_channels>' => 'Yeah, love description',
+                        '<all_channels>' => [
+                            '<all_locales>' => 'Yeah, love description',
                         ],
                     ],
                 ],
@@ -240,8 +240,8 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'identifier' => 'product_4',
                 'values'     => [
                     'description-text' => [
-                        '<all_locales>' => [
-                            '<all_channels>' => 'A better <h1>description</h1>',
+                        '<all_channels>' => [
+                            '<all_locales>' => 'A better <h1>description</h1>',
                         ],
                     ],
                 ],
@@ -250,8 +250,8 @@ class PimCatalogTextAreaIntegration extends AbstractPimCatalogIntegration
                 'identifier' => 'product_5',
                 'values'     => [
                     'description-text' => [
-                        '<all_locales>' => [
-                            '<all_channels>' => 'And an uppercase DESCRIPTION',
+                        '<all_channels>' => [
+                            '<all_locales>' => 'And an uppercase DESCRIPTION',
                         ],
                     ],
                 ],
